requirements.txt
boto3==1.35.14
botocore==1.35.14
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
db-dtypes==1.3.0
dynamodb-json==1.4.2
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.8
jmespath==1.0.1
numpy==2.1.1
packaging==24.1
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.28.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
simplejson==3.19.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
